Reformat intersections files

+++ src/intersections/hit_objs.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   hit_objs.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By:  <>                                        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/02/21 14:44:48 by                   #+#    #+#             */
+/*   Updated: 2022/02/21 14:44:53 by                  ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 //
 // Created by Hajar Sabir on 2/7/22.
 //
@@ -23,7 +35,7 @@ t_colour	cast_ray(t_list **head, t_ray r, t_data img, t_object obj)
 			t = hit_plane(((t_plane *) elem->content), r);
 		else if (elem->type == 'c')
 			t = hit_cylinder(((t_cylinder *) elem->content), r);
-		if (t > 0 && t < obj.distance_to_light - 1e-5f)
+		if (t > 0 && t < obj.distance_to_light - (float)1e-5)
 		{
 			img.light.brightness = 0;
 			return (get_ray_luminosity(img, obj, r));
